evals: small refactor
