+++ evals/test_cases/code_gen_tests/function_tests.py
@@ -1,4 +1,4 @@
-from evals.eval_types import CodeGenTestCaseCore, TestCase
+from evals.eval_types import CodeGenTestCaseCore, CodeGenTestCase
 from evals.notebook_states import *
 
 MY_SUM_FUNCTION = CodeGenTestCaseCore(
@@ -44,12 +44,12 @@ y = sum(x)
 )
 
 FUNCTION_TESTS = [
-    TestCase(
+    CodeGenTestCase(
         name="empty_notebook_function_declaration",
         test_case_core=MY_SUM_FUNCTION,
         user_input="create a function my_sum that takes two arguments and returns their sum. Then use it to create a variable called sum_result that is the sum of 1 and 2",
     ),
-    TestCase(
+    CodeGenTestCase(
         name="palindrome_function",
         test_case_core=PALINDROME_FUNCTION,
         user_input="""Create a function is_palindrome that takes a string and returns True if it is a palindrome and False otherwise. 
@@ -57,7 +57,7 @@ FUNCTION_TESTS = [
 Then use it to create a variable called is_palindrome_result that is True if 'racecar' is a palindrome and False otherwise.
 """,
     ),
-    TestCase(
+    CodeGenTestCase(
         name="fibonacci_function",
         test_case_core=FIBONACCI_FUNCTION,
         user_input="""Create a function fibonacci that takes an integer n and returns an array of the first n Fibonacci numbers where the first two numbers are 1 and 3.
